refactor: move LineHeightAttribute class to custom_attributes.dart
